% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados1.R
\docType{data}
\name{Dados1}
\alias{Dados1}
\title{Dados de exemplo de um experimento sem repeticoes.}
\format{
Um data frame com duas variaveis explicativas quantitativas
 e uma variavel resposta quantitativa.
}
\usage{
data(Dados1)
}
\description{
Exemplo com um conjunto de dados
de um experimento sem delineamento estatistico.Neste caso,
a primeira e segunda coluna devem ser as variaveis explicativas quantiativas
 e a terceira coluna a variavel
resposta quantitativa.
}
\references{
AZEVEDO, AM; ANDRADE JUNIOR, VC; SOUSA JUNIOR, AS; SANTOS, AA; CRUZ, CD;
 PEREIRA, SL; OLIVEIRA, AJM. 2017. Eficiencia da estimacao
 da area foliar de couve por meio de redes neurais artificiais.
  Horticultura Brasileira 35: 014-019.
  DOI - http://dx.doi.org/10.1590/S0102-053620170103

<https://www.youtube.com/playlist?list=PLvth1ZcREyK6OUnWVs-hnyVdCB1xuxbVs>
}
\author{
Alcinei Mistico Azevedo, \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
