% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados2.R
\docType{data}
\name{Dados2}
\alias{Dados2}
\title{Dados de exemplo de um experimento sem repeticoes.}
\format{
Um data frame com duas variaveis explicativas quantitativas
 e uma variavel resposta quantitativa.
}
\usage{
data(Dados2)
}
\description{
Exemplo com um conjunto de dados
de um experimento sem delineamento estatistico.Neste caso, a primeira
 e segunda coluna devem ser as variaveis explicativas quantiativas e a
  terceira coluna a variavel resposta quantitativa.
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK6OUnWVs-hnyVdCB1xuxbVs>
}
\author{
Alcinei Mistico Azevedo, \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
